requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
pyyaml>=6.0

# Vision models
timm>=0.6.0

# NLP metrics
nltk>=3.7
rouge-score>=0.1.2

# Optional: for advanced features
# scikit-learn>=1.0.0
# scipy>=1.7.0

# Logging
tensorboard>=2.10.0
# wandb>=0.13.0  # Optional

# Development
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0


# UniSCC Model Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Vision backbone (Swin Transformer)
timm>=0.9.0

# Text embeddings (CLIP initialization)
transformers>=4.30.0

# Optional dependencies for training
# numpy>=1.24.0
# pillow>=9.5.0
# tqdm>=4.65.0
# tensorboard>=2.13.0
# pyyaml>=6.0

# Development dependencies
# pytest>=7.3.0
# jupyter>=1.0.0
# matplotlib>=3.7.0
